Refactoring backend FastAPI architecture
