+++ backend/.env.sample
@@ -1,0 +1,2 @@
+DATABASE_URL=sqlite:///./app.db
+PROJECT_NAME=KUTechnest API
