+++ backend/crud/student.py
@@ -1,0 +1,16 @@
+from sqlalchemy.orm import Session
+from ..models.student import Student
+from ..schemas.student import StudentCreate
+
+def get_students(db: Session):
+    return db.query(Student).all()
+
+def get_student(db: Session, student_id: int):
+    return db.query(Student).filter(Student.id == student_id).first()
+
+def create_student(db: Session, student: StudentCreate):
+    db_student = Student(**student.dict())
+    db.add(db_student)
+    db.commit()
+    db.refresh(db_student)
+    return db_student
