+++ backend/models/__init__.py
@@ -1,0 +1,6 @@
+from ..models.user import User
+from ..models.company import Company
+from ..models.student import Student
+from ..models.post import Post
+
+__all__ = ["User", "Company", "Student", "Post"]
