+++ backend/schemas/__init__.py
@@ -1,0 +1,9 @@
+from ..schemas.company import CompanyBase, CompanyCreate, CompanyResponse
+from ..schemas.student import StudentBase, StudentCreate, StudentResponse
+from ..schemas.post import PostBase, PostCreate, PostResponse
+
+__all__ = [
+    "CompanyBase", "CompanyCreate", "CompanyResponse",
+    "StudentBase", "StudentCreate", "StudentResponse",
+    "PostBase", "PostCreate", "PostResponse"
+]
